Uninted_integer_division

+++ libraries/AP_HAL/UARTDriver.cpp
@@ -218,9 +218,9 @@ void AP_HAL::UARTDriver::log_stats(const uint8_t inst, StatsTracker &stats, cons
         LOG_PACKET_HEADER_INIT(LOG_UART_MSG),
         time_us  : AP_HAL::micros64(),
         instance : inst,
-        tx_rate  : float((tx_bytes * 1000) / dt_ms),
-        rx_rate  : float((rx_bytes * 1000) / dt_ms),
-        rx_drop_rate : float((rx_dropped_bytes * 1000) / dt_ms),
+        tx_rate  : float((static_cast<float>(tx_bytes) * 1000.0f) / static_cast<float>(dt_ms)),
+        rx_rate  : float((static_cast<float>(rx_bytes) * 1000.0f) / static_cast<float>(dt_ms)), // for uninted integer division error
+        rx_drop_rate : float((static_cast<float>(rx_dropped_bytes) * 1000.0f) / static_cast<float>(dt_ms)),
     };
     AP::logger().WriteBlock(&pkt, sizeof(pkt));
 }
